packages/ai_web_feeds/pyproject.toml
[project]
name = "ai_web_feeds"
version = "0.1.0"
description = "Core library for AI Web Feeds"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.15",
    "alembic>=1.14.0",
    "apscheduler>=3.10.0",
    "beautifulsoup4>=4.12.0",
    "croniter>=5.0.1",
    "feedparser>=6.0.11",
    "gensim>=4.3.0",
    "httpx>=0.28.1",
    "jsonschema>=4.23.0",
    "langdetect>=1.0.9",
    "loguru>=0.7.3",
    "numpy>=2.2.4",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.11.0",
    "python-Levenshtein>=0.26.1",
    "python-socketio>=5.12.0",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.4.0",
    "spacy>=3.7.0",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.27",
    "tenacity>=9.1.2",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "S", "B", "A", "C4", "T10", "T20", "Q", "RET", "SIM", "ARG", "PTH", "PD", "PL", "NPY", "RUF", "ASYNC"]
ignore = ["ANN101", "ANN102", "ANN401"]

[tool.ruff.lint.per-file-ignores]
# Utils: OPML generation uses datetime.utcnow() for RFC compliance
"src/ai_web_feeds/utils.py" = ["DTZ003", "S318", "E501", "RET504"]
# Validate: Legitimate patterns for validation module
"src/ai_web_feeds/validate.py" = ["ANN204", "FBT001", "FBT002", "A001", "ASYNC109", "DTZ005", "PLR2004", "BLE001"]
# WebSocket: Server patterns need specific exceptions
"src/ai_web_feeds/websocket_server.py" = ["ANN204", "S104", "ARG002", "SLF001"]
# Alembic: Generated migration files
"**/alembic/**" = ["INP001", "ERA001", "PLR0915"]
